Network generator tests (stackrox/rox#1527)

+++ qa-tests-backend/src/test/groovy/RbacAuthTest.groovy
@@ -1,16 +1,20 @@
 import groups.BAT
 import io.grpc.StatusRuntimeException
 import io.stackrox.proto.api.v1.AuthproviderService
+import orchestratormanager.OrchestratorTypes
+import org.junit.Assume
 import org.junit.experimental.categories.Category
 import services.ApiTokenService
 import services.AuthProviderService
 import services.BaseService
 import services.ClusterService
+import services.NetworkPolicyService
 import services.RoleService
 import io.stackrox.proto.api.v1.ApiTokenService.GenerateTokenResponse
 import io.stackrox.proto.storage.RoleOuterClass
 import spock.lang.Shared
 import spock.lang.Unroll
+import util.Env
 
 class RbacAuthTest extends BaseSpecification {
     // Create a map of resource name -> functions to execute in read or write scenarios
@@ -28,6 +32,13 @@ class RbacAuthTest extends BaseSpecification {
                                         "stackrox/main:latest",
                                         "central.stackrox:443")
                                 })],
+            "NetworkPolicy": [(RoleOuterClass.Access.READ_ACCESS):
+                                      ( { NetworkPolicyService.generateNetworkPolicies() }),
+                              (RoleOuterClass.Access.READ_WRITE_ACCESS):
+                                      ( {
+                                          NetworkPolicyService.applyGeneratedNetworkPolicy(
+                                                  NetworkPolicyService.generateNetworkPolicies())
+                                      })],
     ]
 
     @Shared
@@ -50,20 +61,22 @@ class RbacAuthTest extends BaseSpecification {
     }
 
     def canDo(Closure closure, String token) {
-        try {
-            BaseService.useApiToken(token)
-            closure()
-        } catch (StatusRuntimeException sre) {
-            return false
-        } finally {
-            BaseService.useBasicAuth()
-        }
-        return true
+        BaseService.useApiToken(token)
+        def result = closure()
+        BaseService.useBasicAuth()
+        return !(result instanceof StatusRuntimeException)
     }
 
     @Unroll
     @Category(BAT)
     def "Verify RBAC with Role/Token combinations: #resourceAccess"() {
+        //skip on OS for now
+        def skip = false
+        if (resourceTest.contains("NetworkPolicy")) {
+            skip = Env.mustGetOrchestratorType() == OrchestratorTypes.OPENSHIFT
+            Assume.assumeFalse(skip)
+        }
+
         when:
         "Create a test role"
         def testRole = RoleOuterClass.Role.newBuilder()
@@ -81,17 +94,30 @@ class RbacAuthTest extends BaseSpecification {
 
         then:
         "verify RBAC permissions"
-        for (String resource : RESOURCE_FUNCTION_MAP.keySet()) {
+        for (String resource : resourceTest) {
             def readFunction = RESOURCE_FUNCTION_MAP.get(resource).get(RoleOuterClass.Access.READ_ACCESS)
             def writeFunction = RESOURCE_FUNCTION_MAP.get(resource).get(RoleOuterClass.Access.READ_WRITE_ACCESS)
-            assert hasReadAccess(resource, resourceAccess) == canDo(readFunction, token.token)
-            assert hasWriteAccess(resource, resourceAccess) == canDo(writeFunction, token.token)
+            println "Testing read function for ${resource}"
+            def read = hasReadAccess(resource, resourceAccess) == canDo(readFunction, token.token)
+            assert read
+            println "Testing write function for ${resource}"
+            def write = hasWriteAccess(resource, resourceAccess) == canDo(writeFunction, token.token)
+            assert write
         }
 
         cleanup:
         "remove role and token"
-        RoleService.deleteRole(testRole.name)
-        if (token.metadata?.id != null) {
+        if (resourceAccess.containsKey("NetworkPolicy") &&
+                resourceAccess.get("NetworkPolicy") == RoleOuterClass.Access.READ_WRITE_ACCESS &&
+                !skip) {
+            NetworkPolicyService.applyGeneratedNetworkPolicy(
+                    NetworkPolicyService.undoGeneratedNetworkPolicy().undoModification
+            )
+        }
+        if (testRole?.name != null) {
+            RoleService.deleteRole(testRole.name)
+        }
+        if (token?.metadata?.id != null) {
             ApiTokenService.revokeToken(token.metadata.id)
         }
         def testClusterId = ClusterService.getClusterId("automation")
@@ -102,10 +128,16 @@ class RbacAuthTest extends BaseSpecification {
         where:
         "Data inputs"
 
-        resourceAccess                                               | _
-        [:]                                                          | _
-        ["Cluster": RoleOuterClass.Access.NO_ACCESS]          | _
-        ["Cluster": RoleOuterClass.Access.READ_ACCESS]        | _
-        ["Cluster": RoleOuterClass.Access.READ_WRITE_ACCESS]  | _
+        resourceAccess                                              | resourceTest
+        [:]                                                         | ["Cluster"]
+        ["Cluster": RoleOuterClass.Access.NO_ACCESS]                | ["Cluster"]
+        ["Cluster": RoleOuterClass.Access.READ_ACCESS]              | ["Cluster"]
+        ["Cluster": RoleOuterClass.Access.READ_WRITE_ACCESS]        | ["Cluster"]
+        ["Cluster": RoleOuterClass.Access.READ_ACCESS,
+         "NetworkGraph": RoleOuterClass.Access.READ_ACCESS,
+         "NetworkPolicy": RoleOuterClass.Access.READ_ACCESS,]       | ["NetworkPolicy"]
+        ["Cluster": RoleOuterClass.Access.READ_ACCESS,
+         "NetworkGraph": RoleOuterClass.Access.READ_ACCESS,
+         "NetworkPolicy": RoleOuterClass.Access.READ_WRITE_ACCESS,] | ["NetworkPolicy"]
     }
 }
